@@ -7,21 +7,20 @@
         "@azure/msal-react": "^1.5.2",
         "@microsoft/applicationinsights-react-js": "^3.4.0",
         "@microsoft/applicationinsights-web": "^2.8.9",
-        "@testing-library/jest-dom": "^5.14.1",
-        "@testing-library/react": "^11.2.7",
-        "@testing-library/user-event": "^12.8.3",
-        "@types/jest": "^26.0.24",
+        "@testing-library/jest-dom": "^5.16.5",
+        "@testing-library/react": "^13.4.0",
+        "@testing-library/user-event": "^14.4.3",
+        "@types/jest": "^29.2.5",
         "@types/node": "^12.20.33",
-        "@types/react": "^17.0.31",
-        "@types/react-dom": "^17.0.10",
-        "bootstrap": "^5.1.3",
-        "react": "^17.0.2",
-        "react-bootstrap": "^2.0.0",
-        "react-dom": "^17.0.2",
-        "react-router-dom": "^5.3.0",
-        "react-scripts": "4.0.3",
-        "typescript": "^4.4.4",
-        "web-vitals": "^1.1.2"
+        "@types/react": "^18.0.26",
+        "@types/react-dom": "^18.0.10",
+        "bootstrap": "^5.2.3",
+        "react": "^18.2.0",
+        "react-bootstrap": "^2.7.0",
+        "react-dom": "^18.2.0",
+        "react-router-dom": "^6.6.2",
+        "typescript": "^4.9.4",
+        "web-vitals": "^3.1.1"
     },
     "scripts": {
         "start": "react-scripts start",
@@ -48,6 +47,7 @@
         ]
     },
     "devDependencies": {
-        "@types/react-router-dom": "^5.3.1"
+        "@types/react-router-dom": "^5.3.3",
+        "react-scripts": "^5.0.1"
     }
 }
